get things sort of working

+++ mysql/tables/Visited.sql
@@ -1,7 +1,7 @@
 USE harvest;
 
 CREATE TABLE IF NOT EXISTS Visited (
-    requestHash INT UNSIGNED NOT NULL,
+    requestHash BIGINT UNSIGNED NOT NULL,
     runId INT NOT NULL,
     PRIMARY KEY (requestHash, runId)
 ) ENGINE=InnoDB;
